requirements.txt
psycopg[binary]==3.2.1
requests==2.32.3
PyYAML==6.0.2
